Remove connect-udp stream type
